+++ Assets/DeltaDNA/Editor/Android/Menus/MenuItems.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright (c) 2016 deltaDNA Ltd. All rights reserved.
+// Copyright (c) 2017 deltaDNA Ltd. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -14,16 +14,50 @@
 // limitations under the License.
 //
 
+using System.IO;
 using UnityEditor;
 using UnityEngine;
 
 namespace DeltaDNA.Editor {
     public sealed class MenuItems : MonoBehaviour {
-        
+
+        internal const string MENU_PATH = "DeltaDNA/Notifications/Android/";
         internal const string EDITOR_PATH = "Assets/DeltaDNA/Editor/";
         
         internal static string AndroidSdkLocation {
             get { return EditorPrefs.GetString("AndroidSdkRoot"); }
+        }
+
+        [MenuItem(MENU_PATH + "Configure", priority = 3)]
+        public static void Configure() {
+            System.Type inspectorType = typeof(UnityEditor.Editor).Assembly.GetType(
+                "UnityEditor.InspectorWindow");
+
+            var foundInspector = false;
+            foreach (var window in Resources.FindObjectsOfTypeAll<EditorWindow>()) {
+                if (window.GetType() == inspectorType) {
+                    foundInspector = true;
+                    break;
+                }
+            }
+
+            if (foundInspector) {
+                EditorWindow.GetWindow<NotificationsWindow>(
+                    "Notifications",
+                    true,
+                    inspectorType)
+                    .Show();
+            } else {
+                EditorWindow.GetWindow<NotificationsWindow>(
+                    "Notifications",
+                    true)
+                    .Show();
+            }
+        }
+
+        internal static bool AreAndroidNotificationsInProject() {
+            return Directory.Exists("Assets/Plugins/Android/deltadna-sdk-unity-notifications")
+                && Directory.Exists("Assets/DeltaDNA/Plugins/Android");
         }
     }
 }